# doc_rank/python-flask-server/swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "Text processing server for upza"
  version: "1.0.0"
  title: "Upza Swagger"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "upza@upza.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/"
schemes:
- "https"
- "http"
paths:
  /actions/process/doc:
    post:
      summary: "Process new doc"
      description: ""
      operationId: "process_doc"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Doc object that needs to be processed"
        required: true
        schema:
          $ref: "#/definitions/ProcessDocRequest"
      responses:
        "200":
          description: "successful operation"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /actions/process/msg:
    post:
      summary: "Process new msg"
      description: ""
      operationId: "process_msg"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Doc object that needs to be processed"
        required: true
        schema:
          $ref: "#/definitions/ProcessMsgRequest"
      responses:
        "200":
          description: "successful operation"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /actions/rank:
    post:
      summary: "Rank Entities"
      description: ""
      operationId: "rank_entities"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Rank arguments"
        required: true
        schema:
          $ref: "#/definitions/RankEntitiesRequest"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RankEntitiesResponse"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /actions/set/cookie:
    post:
      summary: "Set new cookie"
      description: ""
      operationId: "set_cookie"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Cookie that needs to be set"
        required: true
        schema:
          $ref: "#/definitions/SetCookieRequest"
      responses:
        "200":
          description: "successful operation"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
definitions:
  ProcessDocRequest:
    type: "object"
    properties:
      url:
        type: "string"
    example:
      url: "url"
  ProcessMsgRequest:
    type: "object"
    properties:
      msgId:
        type: "string"
      msgText:
        type: "string"
    example:
      msgId: "msgId"
      msgText: "msgText"
  RankEntitiesRequest:
    type: "object"
    properties:
      keywords:
        type: "array"
        items:
          type: "string"
      limit:
        type: "integer"
    example:
      keywords:
      - "keywords"
      - "keywords"
      limit: 0
  SetCookieRequest:
    type: "object"
    properties:
      cookie:
        type: "string"
    example:
      cookie: "s_ppv=[%22www.adobe.com/experience-platform/real-time-customer-profile.html%22%2C20%2C0%2C812%2C1680%2C812%2C1680%2C1050%2C2%2C%22P%22]; JSESSIONID=6F5AF6C8D244947CBBA31730F5D6EAAD"
  RankEntitiesResponse:
    type: "object"
    properties:
      entities:
        type: "array"
        items:
          $ref: "#/definitions/Entity"
    example:
      entities:
      - score: 0
        entityId: "entityId"
      - score: 0
        entityId: "entityId"
  Entity:
    type: "object"
    properties:
      entityId:
        type: "string"
      score:
        type: "integer"
    example:
      score: 0
      entityId: "entityId"